Rename Iris EOL sensor module
