+++ Sample_Test_Script.py
@@ -1,4 +1,4 @@
+import Iris_EOL_Fixture
 
-import Iris_EOL_Sensor_Commands
 if __name__ == "__main__":
-    Iris_EOL_Sensor_Commands.open_monitor()
+    Iris_EOL_Fixture.open_monitor()
